Tested and finished SideBar
