--- src/components/SideBar/SideBarManager/index.js
@@ -1,12 +1,0 @@
-//Imports
-import Routes from '../Routes';
-import SideBarItem from '../SideBarItem';
-import SideBarItems from '../SideBarItems';
-import MainSidebar from '../SideBar';
-
-export {
-    Routes,
-    SideBarItem,
-    SideBarItems,
-    MainSidebar
-}
