+++ src/components/SideBar/index.js
@@ -1,34 +1,16 @@
-/* jshint esversion: 6*/
-import React, { Component } from 'react';
-import { Sidebar, Segment } from 'semantic-ui-react';
-import { HashRouter } from 'react-router-dom';
+//Imports
+import Routes from './Routes';
+import SideBarItem from './SideBarItem';
+import SideBarItems from './SideBarItems';
+import MainSidebar from './SideBarMain';
+import SideBarPusher from './SideBarPusher';
+import SideBar from './SideBar';
 
-import { SideBarItems, Routes, MainSidebar } from './SideBarManager';
-
-import './styles.css';
-
-export default class SideBar extends Component {
-    render() {
-        const SidebarStyle = {
-            borderRadius : 0,
-            borderWidth : 0
-        };
-
-        return (
-            <HashRouter>
-                <div className = "sidebar-container">
-                    <Sidebar.Pushable 
-                        as={ Segment }
-                        style = { SidebarStyle } >
-                        <MainSidebar isVisible = { this.props.visible }>
-                            <SideBarItems toggleVisibility = { this.props.toggleVisibility } />
-                        </MainSidebar>
-                        <Sidebar.Pusher>
-                            <Routes />
-                        </Sidebar.Pusher>
-                    </Sidebar.Pushable>
-                </div>
-            </HashRouter>
-        )
-    }
+export {
+    SideBar,
+    Routes,
+    SideBarItem,
+    SideBarItems,
+    MainSidebar,
+    SideBarPusher
 }
